+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/TestFieldPopulation.kt
@@ -3,20 +3,19 @@ package com.stripe.android
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
-import androidx.test.espresso.IdlingPolicies
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import com.stripe.android.paymentsheet.model.SupportedPaymentMethod
 import com.stripe.android.test.core.Automatic
 import com.stripe.android.test.core.Billing
-import com.stripe.android.test.core.IntentType
 import com.stripe.android.test.core.Currency
 import com.stripe.android.test.core.Customer
 import com.stripe.android.test.core.DelayedPMs
+import com.stripe.android.test.core.MyScreenCaptureProcessor
 import com.stripe.android.test.core.GooglePayState
 import com.stripe.android.test.core.INDIVIDUAL_TEST_TIMEOUT_SECONDS
-import com.stripe.android.test.core.MyScreenCaptureProcessor
+import com.stripe.android.test.core.IntentType
 import com.stripe.android.test.core.PlaygroundTestDriver
 import com.stripe.android.test.core.Shipping
 import com.stripe.android.test.core.TestParameters
@@ -28,7 +27,6 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.Timeout
 import org.junit.runner.RunWith
-import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 class TestFieldPopulation {
@@ -71,9 +69,49 @@ class TestFieldPopulation {
         authorizationAction = null,
     )
 
+    private val card = TestParameters(
+        paymentMethod = SupportedPaymentMethod.Card,
+        customer = Customer.New,
+        googlePayState = GooglePayState.On,
+        currency = Currency.EUR,
+        intentType = IntentType.Pay,
+        billing = Billing.On,
+        shipping = Shipping.Off,
+        delayed = DelayedPMs.On,
+        automatic = Automatic.Off,
+        saveCheckboxValue = false,
+        saveForFutureUseCheckboxVisible = true,
+        useBrowser = null,
+        authorizationAction = null
+    )
+
     @Ignore("Testing of dropdowns is not yet supported")
     fun testDropdowns() {
 
+    }
+
+    @Test
+    fun testCardDefaultAddress() {
+        testDriver.confirmNewOrGuestComplete(
+            card.copy(
+                billing = Billing.On,
+                authorizationAction = null,
+                saveForFutureUseCheckboxVisible = true,
+                saveCheckboxValue = true,
+            )
+        )
+    }
+
+    @Test
+    fun testCardNoDefaultAddress() {
+        testDriver.confirmNewOrGuestComplete(
+            card.copy(
+                billing = Billing.Off,
+                authorizationAction = null,
+                saveForFutureUseCheckboxVisible = true,
+                saveCheckboxValue = false,
+            )
+        )
     }
 
     @Test